Refactor pod model interface
Recenter the implementation on a dataset object which
logs snapshops and keeps inventory records
Add dummy perception model
Paste diff pod model back in
Replumb bin_bounds data out to the outter edge of the node
