+++ aurmr_unseen_object_clustering/src/aurmr_unseen_object_clustering/tools/segmentation_net.py
@@ -26,10 +26,6 @@ import pickle
 
 from aurmr_setup.utils.workspace_utils import get_active_workspace_path
 
-
-rectangle = None
-
-
 workspace_path = get_active_workspace_path()
 UOC_PATH = f'{workspace_path}/src/aurmr_tahoma/aurmr_unseen_object_clustering/src/aurmr_unseen_object_clustering/'
 
@@ -53,12 +49,10 @@ colors_list = ['red', 'green', 'blue', 'magenta', 'yellow', 'cyan', 'purple']
 PIXEL_MEANS = np.array([[[102.9801, 115.9465, 122.7717]]])
 
 import warnings
-warnings.filterwarnings("ignore", category=UserWarning) 
+warnings.filterwarnings("ignore", category=UserWarning)
 
-with open('/tmp/calibration_pixel_coords_pod.pkl', 'rb') as f:
-    bin_bounds = pickle.load(f)
 
-print(bin_bounds)
+
 
 def_config = {
     # Model Parameters
@@ -76,12 +70,11 @@ def_config = {
 #    'model_ref':None,
 
     # Bin / setup parameters
-   'bounds':bin_bounds,
    'image_path':None,
    'min_pixel_threshhold':30,
    'min_match_count': 3,
 
-   # Segmentation / 
+   # Segmentation /
     # Erosion/Dilation and largest connected component
    'perform_cv_ops':False,
    'perform_cv_ops_ref':True,
@@ -113,20 +106,20 @@ class SegNet:
         cfg_from_file(config['cfg_init'])
         if len(cfg.TEST.CLASSES) == 0:
             cfg.TEST.CLASSES = cfg.TRAIN.CLASSES
-       
+
         self.visualize = False
         # Locates the device
         cfg.gpu_id = 0
         cfg.device = torch.device('cuda:{:d}'.format(cfg['gpu_id']))
         cfg.instance_id = 0
-        num_classes = 2 
+        num_classes = 2
         cfg.MODE = 'TEST'
-  
+
         self.config = config
         self.uois_multi_frame = UOISMultiFrame(config["uois_config"], config["uois_weights"])
-        
+
         self.const = 0
-  
+
         # Loads the network for the initial embedding
         # network_data = torch.load(config['model_init'])
         # self.network = networks.__dict__[config['network_name']](num_classes, cfg.TRAIN.NUM_UNITS, network_data).cuda(device=cfg.device)
@@ -146,17 +139,17 @@ class SegNet:
             pass
         else:
             self.network_crop = None
- 
+
          # Turns the initial scene into a point cloud
- 
+
         if init_depth is None:
             print("WARNING: No initial scene provided")
 
         self.H, self.W = None, None
-  
+
         # Contains mappings from pruduct bin ID to [bin, intra_bin_ID]
         self.items = {}
-  
+
         self.n = 0
 
         self.init_depth = init_depth
@@ -167,32 +160,17 @@ class SegNet:
         # List of all bins to ignore
         self.bad_bins = []
 
-  
-   # Computes the point cloud from a depth array and camera intrinsics
-    def compute_xyz(self, depth_img, intrinsic):
-        fx = intrinsic[0][0]
-        fy = intrinsic[1][1]
-        px = intrinsic[0][2]
-        py = intrinsic[1][2]
-        height, width = depth_img.shape
-  
-        indices = np.indices((height, width), dtype=np.float32).transpose(1,2,0)
-        z_e = depth_img
-        x_e = (indices[..., 1] - px) * z_e / fx
-        y_e = (indices[..., 0] - py) * z_e / fy
-        xyz_img = np.stack([x_e, y_e, z_e], axis=-1) # Shape: [H x W x 3]
-        return xyz_img
-    
+
     # Segments the current bin and returns the masks both from the bin reference and from the overall reference
     # Pass in sequence of rgb, xyz pairs already cropped to bin bounds
     def segment(self, rgb_d_sequence: List[Tuple[np.ndarray, np.ndarray]]):
-       
+
         rgb_sequence = [pair[0] for pair in rgb_d_sequence]
         rgb, xyz = rgb_d_sequence[-1]
         # rgb = rgb.astype(np.float32)
 
         # rgb_n = rgb.copy()
-  
+
         # Standardize image and point cloud
         # if self.config['rgb_is_standardized']:
         #     rgb[..., 0] = (rgb[..., 0] - np.mean(rgb[...,0])) / np.std(rgb[...,0])
@@ -206,10 +184,10 @@ class SegNet:
         if self.config['rm_back_old']:
             sd_x, sd_y = self.config['sd_loc']
             sd = xyz[sd_x, sd_y, 2]
-       
+
         # if self.config['resize']:
         #     rgb = cv2.resize(rgb, (256, 256), interpolation=cv2.INTER_AREA)
-        #     xyz = cv2.resize(xyz, (256, 256), interpolation=cv2.INTER_AREA) 
+        #     xyz = cv2.resize(xyz, (256, 256), interpolation=cv2.INTER_AREA)
         #     bg_mask_now = cv2.resize(bin.bg_mask.astype(np.uint8), (256, 256), interpolation=cv2.INTER_AREA)
 
         rgb_segment = rgb.astype(np.uint8)
@@ -240,7 +218,7 @@ class SegNet:
         # print(embed)
         if(np.max(mask) > 0):
             embed = embed.reshape(-1, 256)
-        
+
         out_label = mask
         out_label_new = out_label.astype(np.uint8)
         # out_label_new = out_label\
@@ -253,16 +231,16 @@ class SegNet:
             kernel = np.ones(shape=(self.config['kernel_size'], self.config['kernel_size']), dtype=np.uint8)
             for i in range(np.max(out_label_new.astype(np.uint8))):
                 mask_now = (out_label_new == (i + 1))
-  
+
                 mask_now = spy.binary_erosion(mask_now, structure=kernel, iterations=self.config['erosion_num'])
                 mask_now = spy.binary_dilation(mask_now, structure=kernel, iterations=self.config['dilation_num'])
-  
+
                 labels = lab(mask_now)
 
                 print("labels in perform cv ops", labels)
                 if len(np.bincount(labels.flat)[1:]) > 0:
                     mask_now = (labels == (np.argmax(np.bincount(labels.flat)[1:]) + 1))
-               
+
                 out_label_new[out_label_new == (i + 1)] =    0
                 out_label_new[mask_now] = i + 1
             out_label = torch.reshape(torch.from_numpy(out_label_new).float(), shape=(1, out_label_new.shape[0], out_label_new.shape[1]))
@@ -277,13 +255,13 @@ class SegNet:
             mask_crop_f[mask_crop == index] = i
 
         mask_crop = mask_crop_f
-        
+
         #FIXME COMMENTING BECAUSE MASK ALREADY SEEMS T OBE THE RIGHT SIZE
         #mask_crop = cv2.resize(mask_crop, (bin.bounds[3] - bin.bounds[2], bin.bounds[1] - bin.bounds[0]), interpolation=cv2.INTER_AREA)
 
 
         return mask_crop, embed
-  
+
     # Returns a best-guess matching between masks in frame 0 and masks in frame 1
     #       Returns a list, where the j - 1th element stores the recommended frame 1 ID of the mask with frame 0 ID j
     def match_masks_using_embeddings(self, im1, im2, mask1, mask2, embed1, embed2):
@@ -294,16 +272,12 @@ class SegNet:
         #     pass
         # Convert the images to greyscale
 
-        cv2.imwrite("/home/aurmr/workspaces/aurmr_demo_perception/src/segnetv2_mask2_former/Mask_Results/match_rgb1.png", im1)
-        cv2.imwrite("/home/aurmr/workspaces/aurmr_demo_perception/src/segnetv2_mask2_former/Mask_Results/match_rgb2.png", im2)
-        cv2.imwrite("/home/aurmr/workspaces/aurmr_demo_perception/src/segnetv2_mask2_former/Mask_Results/match_mask1.png", mask1)
-        cv2.imwrite("/home/aurmr/workspaces/aurmr_demo_perception/src/segnetv2_mask2_former/Mask_Results/match_mask2.png", mask2)
         mask_recs = np.zeros(shape=(np.max(mask1), np.max(mask2)))
 
         for i in range(np.max(mask1)):
             max_score = 0
             for j in range(np.max(mask2)):
-                if((np.count_nonzero(embed2[j]) == 0) or (np.count_nonzero(embed1[i]) == 0)): 
+                if((np.count_nonzero(embed2[j]) == 0) or (np.count_nonzero(embed1[i]) == 0)):
                     score = -1
                 else:
                     score = np.dot(embed2[j], embed1[i])
@@ -320,37 +294,35 @@ class SegNet:
 
         print("#############################################")
         print("mask recs", mask_recs)
-  
+
         # Find the mask in the destination image that best matches the soruce mask using Hungarian matching
         row_ind, col_ind = linear_sum_assignment(-mask_recs)
-        self.const 
-  
+        self.const
+
         return col_ind + 1, match_failed, row_ind + 1
-    
+
     def match_masks_using_sift(self, im1, im2, mask1, mask2):
         sift_failed = False
         # Convert the images to greyscale
         im1 = cv2.cvtColor(im1, cv2.COLOR_RGB2GRAY)
         im2 = cv2.cvtColor(im2, cv2.COLOR_RGB2GRAY)
 
-        cv2.imwrite("/home/aurmr/workspaces/aurmr_demo_perception/src/segnetv2_mask2_former/Mask_Results/match_mask1.png", mask1)
-        cv2.imwrite("/home/aurmr/workspaces/aurmr_demo_perception/src/segnetv2_mask2_former/Mask_Results/match_mask2.png", mask2)
         mask_recs = np.zeros(shape=(np.max(mask1), np.max(mask2)))
-  
+
         # Calculate keypoints for the frame 1 image
         #sift = cv2.SIFT_create()
         sift = cv2.SIFT_create(nfeatures=150, contrastThreshold=0.04, edgeThreshold=40)
         # sift = cv2.SIFT_create(nfeatures=150)
         k2, d2 = sift.detectAndCompute(im2, None)
-  
+
         # For each mask in frame 0
         for i in range(1, np.max(mask1) + 1):
             # Subset the image from the mask
             im1_now = im1 * (mask1 == i)
-  
+
             # Calculate keypoints for the masked object in frame 0
             k1, d1 = sift.detectAndCompute(im1_now, None)
-  
+
             # Match keypoints between frames 0 and 1
             FLANN_INDEX_KDTREE = 1
             index_params = dict(algorithm=FLANN_INDEX_KDTREE, trees=5)
@@ -369,7 +341,7 @@ class SegNet:
 
             print(self.config['min_match_count'])
             print(f'found {len(good)} sift matches for object {i}. failed={sift_failed}')
-            
+
 
             # Draws matches for visualization purposes
             matched_img = cv2.drawMatches(im1_now, k1, im2, k2, good, im2, flags=2, matchesThickness=1)
@@ -379,10 +351,10 @@ class SegNet:
             time_stamp = calendar.timegm(current_GMT)
 
             cv2.imwrite(f"/home/aurmr/workspaces/aurmr_demo_perception/src/segnetv2_mask2_former/Mask_Results/matched_mask_{i}.png", matched_img)
-           
+
             dst_pts = np.float32([k2[m.trainIdx].pt for m in good]).reshape(-1, 1, 2)
             dst_pts = np.round(dst_pts).astype(int)
-  
+
             # For each mask in frame 1, find out the number of matched keypoints and add it to n_hits
             for j in range(1, np.max(mask2) + 1):
                 if dst_pts is not None:
@@ -393,10 +365,10 @@ class SegNet:
         print("mask_recs sift", mask_recs)
         # Find the mask in the destination image that best matches the soruce mask sing Hungarian matching
         row_ind, col_ind = linear_sum_assignment(-mask_recs)
-        self.const 
-  
+        self.const
+
         return col_ind + 1, sift_failed, row_ind + 1
-  
+
     # Takes two masks and their recommended frame 0/1 relationship and returns the matched mask 2
     def update_masks(self, mask2, recs, embeddings):
         # try:
@@ -407,7 +379,7 @@ class SegNet:
         mask2_new = np.zeros(shape=mask2.shape, dtype=np.uint8)
         embeddings_new = np.empty([np.max(mask2), 256], dtype=np.float64)
         # embeddings_new = np.copy(embeddings)
-  
+
         # For each mask in recs
         for i in range(len(recs)):
             # Find the mask in the result image
@@ -421,7 +393,7 @@ class SegNet:
         # except:
         #     pass
         return mask2_new, embeddings_new
-  
+
     # Takes a mask and a number of categories and returns a mask that is guaranteed to have that many categories
     def refine_masks(self, mask, n, embed):
         try:
@@ -439,7 +411,7 @@ class SegNet:
         embeddings = np.copy(embed)
         if(n == 0):
             return np.zeros(mask.shape, dtype=np.uint8), np.array([])
-        
+
         # heck if do we need to eliminate any masks
         if(np.max(mask) > n):
             # for every extra masks
@@ -513,13 +485,13 @@ class SegNet:
                     embeddings = np.delete(embeddings, smallest_mask_id-1, 0)
                     print("mask not merged thus removed")
 
-        
+
         # If there are too few masks
         #       split the largest in half along the vertical axis
         if np.max(mask) < n:
             print(f"We only see {np.max(mask)} of {n} masks")
             return None, None
-           
+
             # # Split the largest area along idx into a new mask
             # mask[:, idx:][mask[:, idx:] == idx_max] = np.max(mask) + 1
 
@@ -542,50 +514,6 @@ class SegNet:
             mask_color = np.zeros(image.shape, dtype=image.dtype)
             mask_color[mask == i] = COLORS[colors_list[i - 1]]
             image += .7 * mask_color
-        
+
         image = image.astype(np.uint8)
         return image
-
-    def get_obj_mask_bad_bin(self, obj_id):
-        bin_id, id = self.items[obj_id]
-        bin = self.bins[bin_id]
-        # Mask has n ids from 1 to n
-        mask, embeddings = self.segment(bin_id)
-        # Implement code for using matplotlib to select the relevant mask to return
-        mask2vis = self.vis_masks(bin.current['rgb'], mask)
-
-        def onselect(eclick, erelease):
-            global rectangle
-            rectangle = (eclick.xdata, eclick.ydata, erelease.xdata, erelease.ydata)
-            # plt.close()
-    
-        global rectangle
-        rectangle = None
-
-        # while rectangle is None:
-
-            # fig, ax = plt.subplots(figsize=(8, 6))
-            # __rs = RectangleSelector(ax, onselect)
-
-            # ax.set_title('Select a mask')
-            # ax.imshow(mask2vis)
-            # plt.show()
-        
-        print(rectangle)
-        print(rectangle[0])
-        print(rectangle[0].dtype)
-        label = mask[int(rectangle[1]), int(rectangle[0])]
-        print("We think the label is ", label)
-
-        bin_mask_crop = (mask == label)
-
-        mask_full = np.zeros((self.H, self.W), dtype=np.uint8)
-        r1, r2, c1, c2 = bin.bounds
-        # plt.imshow(mask_crop)
-        # plt.title("mask_crop")
-        # plt.show()
-        mask_full[r1:r2, c1:c2] = np.array((bin_mask_crop == id)).astype(np.uint8)
-
-        return  mask_full
-
-
